+++ src-opam/distro.ml
@@ -376,7 +376,7 @@ let os_family_to_string (os : os_family) =
 let opam_repository (os : os_family) =
   match os with
   | `Cygwin | `Linux -> "https://github.com/ocaml/opam-repository.git"
-  | `Windows -> "https://github.com/fdopen/opam-repository-mingw.git#opam2"
+  | `Windows -> "https://github.com/ocaml-opam/opam-repository-mingw.git#sunset"
 
 let personality os_family arch =
   match os_family with